fix eslint errors
