+++ src/index.js
@@ -7,6 +7,7 @@ import clearList from './modules/clearList';
 import renderList from './modules/renderList';
 import editTask from './modules/editTask';
 import updateStatus from './modules/updateStatus';
+import * as dragNdrop from './modules/drag&drop';
 
 // ------------ EVENT LISTENERS ------------
 // TRAGET DOM ELEMENTS
@@ -14,15 +15,13 @@ const addTaskBtn = document.getElementById('add-task-btn');
 const listPlaceholder = document.getElementById('list-placeholder');
 const clearButton = document.getElementById('clear-btn');
 
-
 addTaskBtn.addEventListener('click', Add);
-
 
 // EDIT A TASK
 listPlaceholder.addEventListener('click', (e) => {
   if (e.target.tagName === 'INPUT' && e.target.classList.contains('task-text')) {
     const taskDescription = e.target;
-    let index = taskDescription.parentElement.parentElement.id;
+    const index = taskDescription.parentElement.parentElement.id;
     taskDescription.addEventListener('keydown', (e) => {
       if (e.keyCode === 13 && taskDescription.value !== '') {
         editTask(taskDescription.value, index);
@@ -48,9 +47,9 @@ listPlaceholder.addEventListener('click', (e) => {
   if (e.target.tagName === 'INPUT' && e.target.classList.contains('checkbox')) {
     const checkbox = e.target;
     const taskText = checkbox.nextElementSibling;
-    taskText.classList.toggle("done");
+    taskText.classList.toggle('done');
     const completed = checkbox.checked;
-    let index = checkbox.parentElement.parentElement.id;
+    const index = checkbox.parentElement.parentElement.id;
     updateStatus(index, completed);
   }
 });
@@ -67,44 +66,19 @@ clearButton.addEventListener('click', () => {
   renderList();
 });
 
-
 // DRAG AND DROP
 listPlaceholder.addEventListener('dragstart', (e) => {
-  if (e.target.classList.contains("task-text")) {
-    let index = e.target.parentElement.parentElement.id;
+  if (e.target.classList.contains('task-text')) {
+    const index = e.target.parentElement.parentElement.id;
     localStorage.setItem('startDragIndex', JSON.stringify(index));
   }
-  const dragTargetArr = [...listPlaceholder.getElementsByTagName("li")];
-  dragTargetArr.forEach(target => {
-    target.addEventListener('dragover', dragOver);
-    target.addEventListener("drop", dragDrop);
+  const dragTargetArr = [...listPlaceholder.getElementsByTagName('li')];
+  dragTargetArr.forEach((target) => {
+    target.addEventListener('dragover', dragNdrop.dragOver());
+    target.addEventListener('drop', dragNdrop.dragDrop());
   });
-})
-
-function dragOver(e) {
-  this.classList.add('drag-over');//colocar margin-top: 40px;
-  e.preventDefault();
-}
-
-function dragDrop() {
-  const dragEndIndex = this.id;
-  const dragStartIndex = JSON.parse(localStorage.getItem('startDragIndex'));
-  console.log("startIndex: ", dragStartIndex);
-  console.log("dropIndex: ", dragEndIndex);
-  swapItems(dragStartIndex, dragEndIndex);
-}
-
-function swapItems(fromIndex, toIndex) {
-  let taskListArr = getArr();
-  taskListArr[fromIndex - 1].Index = toIndex;
-  taskListArr[toIndex - 1].Index = fromIndex;
-  taskListArr.sort((a, b) => a.Index - b.Index);
-  saveInLocalStorage(taskListArr);
-  clearList();
-  renderList();
-}
+});
 
 // SHOW LIST ON HTML
 renderList();
-const focusBackToInput = document.getElementById('task-text').focus();
-
+document.getElementById('task-text').focus();
